Factor out process-message.ts.
